Fix removing card and update form
