--- src/pages/Professor/CreateProfessorPage.tsx
@@ -1,278 +1,0 @@
-import React, { useState } from "react";
-import { useFormik } from "formik";
-import * as Yup from "yup";
-import { Button, Divider, Grid, MenuItem, TextField, Typography } from "@mui/material";
-import { ProfessorInterface } from "../../services/models/Professor";
-import { createProfessor } from "../../services/mentorsService";
-import { FormContainer } from "../CreateGraduation/components/FormContainer";
-import ErrorDialog from "../../components/common/ErrorDialog";
-import SuccessDialog from "../../components/common/SucessDialog";
-import LoadingOverlay from "../../components/common/Loading";
-import {
-  PHONE_ERROR_MESSAGE,
-  PHONE_DIGITS,
-  LETTERS_REGEX,
-  PHONE_REGEX,
-} from "../../constants/validation";
-
-
-const validationSchema = Yup.object({
- name: Yup.string()
-   .matches(LETTERS_REGEX, "El nombre solo debe contener letras")
-   .required("El nombre completo es obligatorio"),
- lastname: Yup.string()
-   .matches(LETTERS_REGEX, "El apellido paterno solo debe contener letras")
-   .required("El apellido es obligatorio"),
- mothername: Yup.string()
-   .matches(LETTERS_REGEX, "El apellido materno solo debe contener letras")
-   .required("El apellido materno es obligatorio"),
- email: Yup.string()
-   .email("Ingrese un correo electrónico válido")
-   .required("El correo electrónico es obligatorio"),
-  phone: Yup.string()
-    .matches(PHONE_REGEX, PHONE_ERROR_MESSAGE)
-    .required("El número de teléfono es requerido"),
- degree: Yup.string().required("El título académico es obligatorio"),
- code: Yup.number()
-   .typeError("El código debe ser numérico")
-   .required("El código de docente es obligatorio"),
-
-});  
-
-
-const CreateProfessorPage = () => {
- const [loading, setLoading] = useState(false);
- const [message, setMessage] = useState("");
- const [successDialog, setSuccessDialog] = useState(false);
- const [errorDialog, setErrorDialog] = useState(false);
-
-
- const sucessDialogClose = () => {
-   setSuccessDialog(false);
-   formik.resetForm();
- };
-
-
- const errorDialogClose = () => {
-   setErrorDialog(false);
- };
-
-
- const formik = useFormik<ProfessorInterface>({
-   initialValues: {
-     name: "",
-     lastname: "",
-     mothername: "",
-     email: "",
-     phone: "",
-     degree: "",
-     code: "",
-   },
-   validationSchema,
-   onSubmit: async (values) => {
-     setLoading(true);
-     try {
-       await createProfessor(values);
-       setMessage("Profesor creado con éxito");
-       setSuccessDialog(true);
-     } catch (error) {
-       setMessage("Error al crear el docente");
-       setErrorDialog(true);
-     } finally {
-       setLoading(false);
-     }
-   },
- });
-
-
-  const handlePhoneChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const { value } = event.target;
-    if (/^[0-9]*$/.test(value)) {
-      formik.setFieldValue("phone", value);
-    }
-  };
-
-
- const handleCodeChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-   const value = event.target.value;
-   if (/^[0-9]*$/.test(value)) {
-     formik.setFieldValue("code", value);
-   }
- };
-
-
- return (
-   <FormContainer>
-     {loading && <LoadingOverlay message="Creando Docente..." />}
-     <form onSubmit={formik.handleSubmit}>
-       <Grid container spacing={2} sx={{ padding: 2 }}>
-         <Grid item xs={12}>
-           <Typography variant="h4">Crear Nuevo Docente</Typography>
-           <Typography variant="body2" sx={{ fontSize: 14, color: "gray" }}>
-             Ingrese los datos del docente a continuación.
-           </Typography>
-           <Divider flexItem sx={{ mt: 2, mb: 2 }} />
-         </Grid>
-
-
-         <Grid item xs={12}>
-           <Grid container spacing={2} sx={{ padding: 2 }}>
-             <Grid item xs={3}>
-               <Typography variant="h6">Información del Docente</Typography>
-             </Grid>
-             <Grid item xs={9}>
-               <Grid container spacing={2}>
-                 <Grid item xs={6}>
-                   <TextField
-                     id="name"
-                     name="name"
-                     label="Nombres"
-                     variant="outlined"
-                     fullWidth
-                     value={formik.values.name}
-                     onChange={formik.handleChange}
-                     error={formik.touched.name && Boolean(formik.errors.name)}
-                     helperText={formik.touched.name && formik.errors.name}
-                     margin="normal"
-                   />
-                 </Grid>
-                 <Grid item xs={6}>
-                   <TextField
-                     id="lastname"
-                     name="lastname"
-                     label="Apellido Paterno"
-                     variant="outlined"
-                     fullWidth
-                     value={formik.values.lastname}
-                     onChange={formik.handleChange}
-                     error={formik.touched.lastname && Boolean(formik.errors.lastname)}
-                     helperText={formik.touched.lastname && formik.errors.lastname}
-                     margin="normal"
-                   />
-                 </Grid>
-               </Grid>
-               <Grid container spacing={2}>
-                 <Grid item xs={6}>
-                   <TextField
-                     id="mothername"
-                     name="mothername"
-                     label="Apellido Materno"
-                     variant="outlined"
-                     fullWidth
-                     value={formik.values.mothername}
-                     onChange={formik.handleChange}
-                     error={formik.touched.mothername &&
-Boolean(formik.errors.mothername)}
-                     helperText={formik.touched.mothername && formik.errors.mothername}
-                     margin="normal"
-                   />
-                 </Grid>
-                 <Grid item xs={6}>
-                   <TextField
-                     id="code"
-                     name="code"
-                     label="Codigo de Docente"
-                     variant="outlined"
-                     fullWidth
-                     value={formik.values.code}
-                     onChange={handleCodeChange}
-                     error={formik.touched.code && Boolean(formik.errors.code)}
-                     helperText={formik.touched.code && formik.errors.code}
-                     margin="normal"
-                     inputProps={{ maxLength: 10 }}
-                   />
-                 </Grid>
-               </Grid>
-               <TextField
-                 id="degree"
-                 name="degree"
-                 label="Título Académico"
-                 variant="outlined"
-                 fullWidth
-                 select
-                 value={formik.values.degree}
-                 onChange={formik.handleChange}
-                 error={formik.touched.degree && Boolean(formik.errors.degree)}
-                 helperText={formik.touched.degree && formik.errors.degree}
-                 margin="normal"
-               >
-                 <MenuItem value="">Seleccione un título</MenuItem>
-                 <MenuItem value="Ing.">Ing.</MenuItem>
-                 <MenuItem value="M.Sc.">M.Sc.</MenuItem>
-                 <MenuItem value="PhD.">PhD.</MenuItem>
-                 <MenuItem value="M.Eng.">M.Eng.</MenuItem>
-               </TextField>
-             </Grid>
-           </Grid>
-           <Divider flexItem sx={{ my: 2 }} />
-         </Grid>
-
-
-         <Grid item xs={12}>
-           <Grid container spacing={2} sx={{ padding: 2 }}>
-             <Grid item xs={3}>
-               <Typography variant="h6">Información Adicional</Typography>
-             </Grid>
-             <Grid item xs={9}>
-               <TextField
-                 id="email"
-                 name="email"
-                 label="Correo Electrónico"
-                 variant="outlined"
-                 fullWidth
-                 value={formik.values.email}
-                 onChange={formik.handleChange}
-                 error={formik.touched.email && Boolean(formik.errors.email)}
-                 helperText={formik.touched.email && formik.errors.email}
-                 margin="normal"
-                 inputProps={{ maxLength: 50 }}
-               />
-               <TextField
-                 id="phone"
-                 name="phone"
-                 label="Número de Teléfono"
-                 variant="outlined"
-                 fullWidth
-                 value={formik.values.phone}
-                 onChange={handlePhoneChange}
-                 error={formik.touched.phone && Boolean(formik.errors.phone)}
-                 helperText={formik.touched.phone && formik.errors.phone}
-                 margin="normal"
-                  inputProps={{ maxLength: PHONE_DIGITS }}
-                />
-             </Grid>
-           </Grid>
-         </Grid>
-         <Grid item xs={12}>
-           <Grid container spacing={2} justifyContent="flex-end">
-             <Grid item>
-               <Button variant="contained" color="primary" type="submit" sx={{ mr: 3 }}>
-                 GUARDAR
-               </Button>
-               <Button variant="contained" color="primary" type="button"
-               onClick={() => window.history.back()}>
-                 CANCELAR
-               </Button>
-             </Grid>
-           </Grid>
-         </Grid>
-       </Grid>
-     </form>
-     <SuccessDialog
-       open={successDialog}
-       onClose={sucessDialogClose}
-       title={"Docente Creado!"}
-       subtitle={"El docente ha sido creado con éxito."}
-     />
-     <ErrorDialog
-       open={errorDialog}
-       onClose={errorDialogClose}
-       title={"¡Vaya!"}
-       subtitle={message}
-     />
-   </FormContainer>
- );
-};
-
-
-export default CreateProfessorPage;
